+++ Microsoft.Azure.Cosmos/src/CosmosClientOptions.cs
@@ -341,6 +341,16 @@ namespace Microsoft.Azure.Cosmos
         }
 
         /// <summary>
+        /// Allows optimistic batching of requests to service. Setting this option might impact the latency of the operations. Hence this option is recommended for non-latency sensitive scenarios only.
+        /// </summary>
+#if PREVIEW
+        public
+#else
+        internal
+#endif
+        bool AllowBulkExecution { get; set; }
+
+        /// <summary>
         /// A JSON serializer used by the CosmosClient to serialize or de-serialize cosmos request/responses.
         /// The default serializer is always used for all system owned types like DatabaseProperties.
         /// The default serializer is used for user types if no UserJsonSerializer is specified