+++ Microsoft.Azure.Cosmos/src/Resource/Offer/CosmosOffers.cs
@@ -10,6 +10,7 @@ namespace Microsoft.Azure.Cosmos
     using System.Net;
     using System.Threading;
     using System.Threading.Tasks;
+    using Microsoft.Azure.Cosmos.Handlers;
     using Microsoft.Azure.Documents;
 
     internal class CosmosOffers
@@ -28,10 +29,30 @@ namespace Microsoft.Azure.Cosmos
             RequestOptions requestOptions,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            OfferV2 offerV2 = await this.GetOfferV2Async(targetRID, cancellationToken);
+            OfferV2 offerV2 = await this.GetOfferV2Async(targetRID, failIfNotConfigured: true, cancellationToken: cancellationToken);
+
+            return await this.GetThroughputResponseAsync(
+                streamPayload: null,
+                operationType: OperationType.Read,
+                linkUri: new Uri(offerV2.SelfLink, UriKind.Relative),
+                resourceType: ResourceType.Offer,
+                requestOptions: requestOptions,
+                cancellationToken: cancellationToken);
+        }
+
+        internal async Task<ThroughputResponse> ReadThroughputIfExistsAsync(
+            string targetRID,
+            RequestOptions requestOptions,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            OfferV2 offerV2 = await this.GetOfferV2Async(targetRID, failIfNotConfigured: false, cancellationToken: cancellationToken);
+
             if (offerV2 == null)
             {
-                return new ThroughputResponse(HttpStatusCode.NotFound, null, null);
+                return new ThroughputResponse(
+                    HttpStatusCode.NotFound,
+                    headers: null,
+                    throughputProperties: null);
             }
 
             return await this.GetThroughputResponseAsync(
@@ -50,12 +71,7 @@ namespace Microsoft.Azure.Cosmos
             RequestOptions requestOptions,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            OfferV2 offerV2 = await this.GetOfferV2Async(targetRID, cancellationToken);
-            if (offerV2 == null)
-            {
-                throw new CosmosException(HttpStatusCode.NotFound, $"Throughput is not configured for resource:{resourceName}");
-            }
-
+            OfferV2 offerV2 = await this.GetOfferV2Async(targetRID, failIfNotConfigured: true, cancellationToken: cancellationToken);
             OfferV2 newOffer = new OfferV2(offerV2, throughput);
 
             return await this.GetThroughputResponseAsync(
@@ -67,9 +83,47 @@ namespace Microsoft.Azure.Cosmos
                 cancellationToken: cancellationToken);
         }
 
-        internal async Task<OfferV2> GetOfferV2Async(
+        internal async Task<ThroughputResponse> ReplaceThroughputIfExistsAsync(
             string targetRID,
+            int throughput,
+            RequestOptions requestOptions,
             CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                OfferV2 offerV2 = await this.GetOfferV2Async(targetRID, failIfNotConfigured: true, cancellationToken: cancellationToken);
+                OfferV2 newOffer = new OfferV2(offerV2, throughput);
+
+                return await this.GetThroughputResponseAsync(
+                    streamPayload: this.ClientContext.PropertiesSerializer.ToStream(newOffer),
+                    operationType: OperationType.Replace,
+                    linkUri: new Uri(offerV2.SelfLink, UriKind.Relative),
+                    resourceType: ResourceType.Offer,
+                    requestOptions: requestOptions,
+                    cancellationToken: cancellationToken);
+            }
+            catch (DocumentClientException dce)
+            {
+                ResponseMessage responseMessage = dce.ToCosmosResponseMessage(null);
+                return new ThroughputResponse(
+                    responseMessage.StatusCode,
+                    headers: responseMessage.Headers,
+                    throughputProperties: null);
+            }
+            catch (AggregateException ex)
+            {
+                ResponseMessage responseMessage = TransportHandler.AggregateExceptionConverter(ex, null);
+                return new ThroughputResponse(
+                    responseMessage.StatusCode,
+                    headers: responseMessage.Headers,
+                    throughputProperties: null);
+            }
+        }
+
+        private async Task<OfferV2> GetOfferV2Async(
+            string targetRID,
+            bool failIfNotConfigured,
+            CancellationToken cancellationToken)
         {
             if (string.IsNullOrWhiteSpace(targetRID))
             {
@@ -80,17 +134,26 @@ namespace Microsoft.Azure.Cosmos
             queryDefinition.WithParameter("@targetRID", targetRID);
 
             FeedIterator<OfferV2> databaseStreamIterator = this.GetOfferQueryIterator<OfferV2>(
-                 queryDefinition);
+                 queryDefinition: queryDefinition,
+                 continuationToken: null,
+                 requestOptions: null,
+                 cancellationToken: cancellationToken);
             OfferV2 offerV2 = await this.SingleOrDefaultAsync<OfferV2>(databaseStreamIterator);
+
+            if (offerV2 == null &&
+                failIfNotConfigured)
+            {
+                throw new CosmosException(HttpStatusCode.NotFound, "Throughput is not configured");
+            }
 
             return offerV2;
         }
 
         internal virtual FeedIterator<T> GetOfferQueryIterator<T>(
             QueryDefinition queryDefinition,
-            string continuationToken = null,
-            QueryRequestOptions requestOptions = null,
-            CancellationToken cancellationToken = default(CancellationToken))
+            string continuationToken,
+            QueryRequestOptions requestOptions,
+            CancellationToken cancellationToken)
         {
             FeedIterator databaseStreamIterator = this.GetOfferQueryStreamIterator(
                queryDefinition,
@@ -117,22 +180,6 @@ namespace Microsoft.Azure.Cosmos
                continuationToken: continuationToken,
                options: requestOptions,
                usePropertySerializer: true);
-        }
-
-        private CosmosOfferResult GetThroughputIfExists(Offer offer)
-        {
-            if (offer == null)
-            {
-                return new CosmosOfferResult(null);
-            }
-
-            OfferV2 offerV2 = offer as OfferV2;
-            if (offerV2 == null)
-            {
-                throw new NotImplementedException();
-            }
-
-            return new CosmosOfferResult(offerV2.Content.OfferThroughput);
         }
 
         private async Task<T> SingleOrDefaultAsync<T>(